Clean up Makefiles and scripts across all repos

+++ scripts/build-dists.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env bash
 set -e
 
+if [[ "$1" == '--no-build' ]]; then
+    no_build='yes'
+fi
+
 # Directory of *this* script
 this_dir="$( cd "$( dirname "$0" )" && pwd )"
 src_dir="$(realpath "${this_dir}/..")"
@@ -12,16 +16,18 @@ src_dir="$(realpath "${this_dir}/..")"
 dist_dir="${src_dir}/dist"
 mkdir -p "${dist_dir}"
 
-# Clear old Rhasspy libraries
-rm -f "${dist_dir}/"rhasspy*
+if [[ -z "${no_build}" ]]; then
+    # Clear old Rhasspy libraries
+    rm -f "${dist_dir}/"rhasspy*
 
-# Make dependent libraries
-while read -r dir_name;
-do
-    echo "${dir_name}"
-    cd "${src_dir}/${dir_name}" && "${PYTHON}" setup.py sdist --dist-dir "${dist_dir}"
-    echo ""
-done < "${src_dir}/RHASSPY_DIRS"
+    # Make dependent libraries
+    while read -r dir_name;
+    do
+        echo "${dir_name}"
+        cd "${src_dir}/${dir_name}" && "${PYTHON}" setup.py sdist --dist-dir "${dist_dir}"
+        echo ""
+    done < "${src_dir}/RHASSPY_DIRS"
+fi
 
 # Update submodule downloads
 while read -r dir_name;
@@ -33,10 +39,21 @@ do
     rm -f "${download_dir}"/rhasspy*.tar.gz
 
     # Copy required libraries
-    while read -r lib_name;
-    do
-        dist_name="$(echo "${lib_name}" | sed -e 's/[=~]=/-/').tar.gz"
-        cp "${dist_dir}/${dist_name}" "${download_dir}/"
-        echo "${dir_name} <- ${dist_name}"
-    done < <(grep '^rhasspy-' "${src_dir}/${dir_name}/requirements.txt")
+    lib_dir_names=("${dir_name}")
+    while [[ -n "${lib_dir_names[@]}" ]]; do
+        # Pop first item
+        lib_dir_name="${lib_dir_names[0]}"
+        lib_dir_names=("${lib_dir_names[@]:1}")
+
+        while read -r lib_name;
+        do
+            dist_name="$(echo "${lib_name}" | sed -e 's/[=~]=/-/').tar.gz"
+            cp "${dist_dir}/${dist_name}" "${download_dir}/"
+            echo "${dir_name} <- ${dist_name}"
+
+            # Search for more Rhasspy dependencies
+            sub_dir_name="$(echo "${lib_name}" | sed -e 's/[=~]=.\+$//')"
+            lib_dir_names+=("${sub_dir_name}")
+        done < <(grep '^rhasspy-' "${src_dir}/${lib_dir_name}/requirements.txt")
+    done
 done < "${src_dir}/RHASSPY_DIRS"
